library/tracker-component-refactoring
Created constant file to save mock data, interfaces, etc
