+++ src/components/tracker/Tracker.stories.ts
@@ -1,5 +1,6 @@
 import type { Meta, StoryObj } from "@storybook/react";
 import { DefaultTracker as Tracker } from "./Tracker";
+import { data } from "./constants";
 
 const meta = {
   title: "Taggify/Tracker/Tracker",
@@ -17,32 +18,6 @@ type Story = StoryObj<typeof meta>;
 // More on writing stories with args: https://storybook.js.org/docs/react/writing-stories/args
 export const DefaultTracker: Story = {
   args: {
-    data: [
-      { color: "emerald", tooltip: "Operational" },
-      { color: "emerald", tooltip: "Operational" },
-      { color: "emerald", tooltip: "Operational" },
-      { color: "emerald", tooltip: "Operational" },
-      { color: "emerald", tooltip: "Operational" },
-      { color: "emerald", tooltip: "Operational" },
-      { color: "emerald", tooltip: "Operational" },
-      { color: "emerald", tooltip: "Operational" },
-      { color: "emerald", tooltip: "Operational" },
-      { color: "emerald", tooltip: "Operational" },
-      { color: "emerald", tooltip: "Operational" },
-      { color: "emerald", tooltip: "Operational" },
-      { color: "emerald", tooltip: "Operational" },
-      { color: "emerald", tooltip: "Operational" },
-      { color: "rose", tooltip: "Downtime" },
-      { color: "emerald", tooltip: "Operational" },
-      { color: "emerald", tooltip: "Operational" },
-      { color: "emerald", tooltip: "Operational" },
-      { color: "emerald", tooltip: "Operational" },
-      { color: "gray", tooltip: "Maintenance" },
-      { color: "emerald", tooltip: "Operational" },
-      { color: "emerald", tooltip: "Operational" },
-      { color: "emerald", tooltip: "Operational" },
-      { color: "yellow", tooltip: "Degraded" },
-      { color: "emerald", tooltip: "Operational" },
-    ],
+    data,
   },
 };
