+++ src/components/tracker/constants.ts
@@ -1,0 +1,38 @@
+import { Color } from "@tremor/react";
+
+export interface DataProps {
+  color: Color;
+  tooltip: string;
+}
+
+export interface TrackerProps {
+  data: DataProps[];
+}
+
+export const data: DataProps[] = [
+  { color: "emerald", tooltip: "Operational" },
+  { color: "emerald", tooltip: "Operational" },
+  { color: "emerald", tooltip: "Operational" },
+  { color: "emerald", tooltip: "Operational" },
+  { color: "emerald", tooltip: "Operational" },
+  { color: "emerald", tooltip: "Operational" },
+  { color: "emerald", tooltip: "Operational" },
+  { color: "emerald", tooltip: "Operational" },
+  { color: "emerald", tooltip: "Operational" },
+  { color: "emerald", tooltip: "Operational" },
+  { color: "emerald", tooltip: "Operational" },
+  { color: "emerald", tooltip: "Operational" },
+  { color: "emerald", tooltip: "Operational" },
+  { color: "emerald", tooltip: "Operational" },
+  { color: "rose", tooltip: "Downtime" },
+  { color: "emerald", tooltip: "Operational" },
+  { color: "emerald", tooltip: "Operational" },
+  { color: "emerald", tooltip: "Operational" },
+  { color: "emerald", tooltip: "Operational" },
+  { color: "gray", tooltip: "Maintenance" },
+  { color: "emerald", tooltip: "Operational" },
+  { color: "emerald", tooltip: "Operational" },
+  { color: "emerald", tooltip: "Operational" },
+  { color: "yellow", tooltip: "Degraded" },
+  { color: "emerald", tooltip: "Operational" },
+];
